fix: ijepa example
