+++ lightly/models/utils.py
@@ -1261,3 +1261,43 @@ def update_drop_path_rate(
         else:
             block.drop_path1 = Identity()
             block.drop_path2 = Identity()
+
+
+def repeat_interleave_batch(x: torch.Tensor, B: int, repeat: int) -> torch.Tensor:
+    """Repeat and interleave the input tensor."""
+    N = len(x) // B
+    x = torch.cat(
+        [
+            torch.cat([x[i * B : (i + 1) * B] for _ in range(repeat)], dim=0)
+            for i in range(N)
+        ],
+        dim=0,
+    )
+    return x
+
+
+def apply_masks(
+    x: torch.Tensor, masks: torch.Tensor | list[torch.Tensor]
+) -> torch.Tensor:
+    """Apply masks to the input tensor.
+
+    From https://github.com/facebookresearch/ijepa/blob/main/src/masks/utils.py
+
+    Args:
+        x:
+            tensor of shape [B (batch-size), N (num-patches), D (feature-dim)].
+        masks:
+            tensor or list of tensors containing indices of patches in [N] to keep.
+
+    Returns:
+        Tensor of shape [B, N', D] where N' is the number of patches to keep.
+    """
+
+    if not isinstance(masks, list):
+        masks = [masks]
+
+    all_x = []
+    for m in masks:
+        mask_keep = m.unsqueeze(-1).repeat(1, 1, x.size(-1))
+        all_x += [torch.gather(x, dim=1, index=mask_keep)]
+    return torch.cat(all_x, dim=0)
